fix migration error up and down

+++ database/migrations/2025_07_11_111454_alter_orders_table.php
@@ -13,6 +13,10 @@ return new class extends Migration
     {
         Schema::table('orders', function (Blueprint $table) {
             $table->unsignedBigInteger('employee_id')->nullable(); // Employee handling the order, if applicable.
+            $table->foreign('employee_id')
+                ->references('employee_id')
+                ->on('employees')
+                ->onDelete('set null'); // Foreign key constraint to employees table, set to null if employee is deleted
         });
     }
 
@@ -22,8 +26,10 @@ return new class extends Migration
     public function down(): void
     {
         Schema::table('orders', function (Blueprint $table) {
-            $table->dropForeign(['employee_id']);
-            $table->dropColumn('employee_id'); // Remove employee_id column
+            if (Schema::hasColumn('orders', 'employee_id')) {
+                $table->dropForeign(['employee_id']); // Drop foreign key constraint
+                $table->dropColumn('employee_id'); // Remove employee_id column
+            }
         });
     }
 };
